+++ lib/core/manhwa_sites/asura_scans.dart
@@ -1,6 +1,7 @@
 import 'dart:developer';
 
 import 'package:manga_reader/core/utils.dart';
+import 'package:manga_reader/core/webscraper_extension.dart';
 import 'package:web_scraper/web_scraper.dart';
 
 import '../core_types.dart';
@@ -10,23 +11,20 @@ class AsuraScans implements ManhwaSource {
 
   @override
   Future<List<String>> getChapterImages(String chapterUrl) async {
-    final chapterRoute = chapterUrl.replaceAll(RegExp('https://www.asurascans.com'), '');
+    final chapterRoute =
+        chapterUrl.replaceAll(RegExp('https://www.asurascans.com'), '');
 
     if (await _webScraper.loadWebPage(chapterRoute)) {
-      List<String> allImages = _webScraper
-          .getElement('div#readerarea > p > img', ['src'])
-          .map(
-            (e) => e['attributes']['src'].toString(),
-          )
-          .toList();
+      List<String> allImages = _webScraper.getElementAttributeUnwrapString(
+          'div#readerarea > p > img', 'src');
 
       // checks if there are anomalous images that are not in <p></p>
-      final anomalousImages =
-          _webScraper.getElement('div#readerarea > img', ['src']);
+      final anomalousImages = _webScraper.getElementAttributeUnwrapString(
+          'div#readerarea > img', 'src');
 
       if (anomalousImages.isNotEmpty) {
         for (var i = 0; i < anomalousImages.length; i++) {
-          final anomalousImage = anomalousImages[i]['attributes']['src'];
+          final anomalousImage = anomalousImages[i];
           final fileName = anomalousImage.toString().split('/').last;
           final index = int.parse(fileName.split('-').first);
 
@@ -50,45 +48,42 @@ class AsuraScans implements ManhwaSource {
     if (await _webScraper.loadWebPage(mangaRoute)) {
       // ─── Get Title ───────────────────────────────────────
       final mangaTitle =
-          _webScraper.getElementTitle('div.infox > h1.entry-title').first;
+          _webScraper.getFirstElementTitle('div.infox > h1.entry-title');
 
       // ─── Get Description ─────────────────────────────────
       final mangaDescription = removeExtraWhiteSpaces(
-        _webScraper.getElementTitle('div.wd-full > div.entry-content').first,
+        _webScraper.getFirstElementTitle('div.wd-full > div.entry-content'),
       );
 
       // ─── Get Cover Url ───────────────────────────────────
-      final mangaCoverUrl = _webScraper
-          .getElement('div.thumbook > div.thumb > img', ['src'])
-          .first['attributes']['src']
-          .toString();
+      final mangaCoverUrl = _webScraper.getFirstElementAttribute(
+          'div.thumbook > div.thumb > img', 'src');
 
       // ─── Get Status ──────────────────────────────────────
       final mangaStatus = MangaStatus.parse(
-        _webScraper.getElementTitle('div.imptdt > i').first,
+        _webScraper.getFirstElementTitle('div.imptdt > i'),
       );
 
       // ─── Get Rating ──────────────────────────────────────
       final mangaRating = double.parse(
-        _webScraper
-            .getElementTitle('div.rt > div.rating > div.rating-prc > div.num')
-            .first,
+        _webScraper.getFirstElementTitle(
+            'div.rt > div.rating > div.rating-prc > div.num'),
       );
 
       // ─── Get Released At ─────────────────────────────────
       final mangaReleasedAt = DateTime.parse(
-        _webScraper.getElement(
-          'div.flex-wrap > div.fmed > span > time',
-          ['datetime'],
-        ).first['attributes']['datetime'],
+        _webScraper.getFirstElementAttribute(
+            'div.flex-wrap > div.fmed > span > time', 'datetime'),
       );
 
       // ─── Get Tags ────────────────────────────────────────
       final mangaTags = _webScraper.getElementTitle('span.mgen > a');
 
       // ─── Get Chapters ────────────────────────────────────
-      final mangaChapterTitles =
-          _webScraper.getElementTitle('span.chapternum').map((e) => removeChapterFromString(e)).toList();
+      final mangaChapterTitles = _webScraper
+          .getElementTitle('span.chapternum')
+          .map((e) => removeChapterFromString(e))
+          .toList();
 
       final mangaChapterReleasedOns = _webScraper
           .getElementTitle('span.chapterdate')
@@ -97,12 +92,8 @@ class AsuraScans implements ManhwaSource {
           )
           .toList();
 
-      final mangaChapterUrls = _webScraper
-          .getElement('div.eph-num > a', ['href'])
-          .map(
-            (e) => e['attributes']['href'].toString(),
-          )
-          .toList();
+      final mangaChapterUrls = _webScraper.getElementAttributeUnwrapString(
+          'div.eph-num > a', 'href');
 
       List<MangaChapterData> mangaChapters = [];
       for (var i = 0; i < mangaChapterUrls.length; i++) {
@@ -134,80 +125,32 @@ class AsuraScans implements ManhwaSource {
 
   @override
   Future<List<MangaSearchResult>> popular({int page = 1}) async {
-    if (await _webScraper
-        .loadWebPage('/manga/?status=&type=&order=popular&page=$page')) {
-      // ─── Get Cover Urls ──────────────────
-      final resultCoverUrls = _webScraper
-          .getElement('div.bsx > a > div.limit > img.ts-post-image', ['src'])
-          .map(
-            (e) => e['attributes']['src'].toString(),
-          )
-          .toList();
+    final targetEndpoint = '/manga/?status=&type=&order=popular&page=$page';
 
-      // ─── Get Titles ──────────────────────
-      final resultTitles = _webScraper.getElementTitle('div.bigor > div.tt');
-
-      // ─── Get Latest Chapter Numbers ──────
-      final resultLatestChapterTitles =
-          _webScraper.getElementTitle('div.bigor > div.adds > div.epxs').map((e) => removeChapterFromString(e)).toList();
-
-      // ─── Get Ratings ─────────────────────
-      final resultRatings = _webScraper
-          .getElementTitle('div.adds > div.rt > div.rating > div.numscore')
-          .map(
-            (e) => double.parse(e),
-          )
-          .toList();
-
-      // ─── Get Manga Urls ──────────────────
-      final resultMangaUrls = _webScraper
-          .getElement('div.bsx > a', ['href'])
-          .map(
-            (e) => e['attributes']['href'].toString(),
-          )
-          .toList();
-
-      // ─── Into Mangasearchresult List ─────
-
-      List<MangaSearchResult> results = [];
-
-      for (var i = 0; i < resultMangaUrls.length; i++) {
-        results.add(
-          MangaSearchResult(
-            resultCoverUrls[i],
-            resultTitles[i],
-            resultLatestChapterTitles[i],
-            resultRatings[i],
-            resultMangaUrls[i],
-            MangaStatus.none,
-          ),
-        );
-      }
-
-      return results;
-    }
-
-    return [];
+    return await _makeSearch(targetEndpoint);
   }
 
   @override
   Future<List<MangaSearchResult>> search(String query) async {
-    final formattedQuery = query.replaceAll(RegExp(' '), '+').toLowerCase();
-    if (await _webScraper.loadWebPage('/?s=$formattedQuery')) {
+    final targetEndpoint = query.replaceAll(RegExp(' '), '+').toLowerCase();
+
+    return await _makeSearch('/?s=$targetEndpoint');
+  }
+
+  Future<List<MangaSearchResult>> _makeSearch(String targetEndpoint) async {
+    if (await _webScraper.loadWebPage(targetEndpoint)) {
       // ─── Get Cover Urls ──────────────────
-      final resultCoverUrls = _webScraper
-          .getElement('div.dsx > a > div.limit > img', ['src'])
-          .map(
-            (e) => e['attributes']['src'].toString(),
-          )
-          .toList();
+      final resultCoverUrls = _webScraper.getElementAttributeUnwrapString(
+          'div.bsx > a > div.limit > img.ts-post-image', 'src');
 
       // ─── Get Titles ──────────────────────
       final resultTitles = _webScraper.getElementTitle('div.bigor > div.tt');
 
       // ─── Get Latest Chapter Numbers ──────
-      final resultLatestChapterTitles =
-          _webScraper.getElementTitle('div.bigor > div.adds > div.epxs').map((e) => removeChapterFromString(e)).toList();
+      final resultLatestChapterTitles = _webScraper
+          .getElementTitle('div.bigor > div.adds > div.epxs')
+          .map((e) => removeChapterFromString(e))
+          .toList();
 
       // ─── Get Ratings ─────────────────────
       final resultRatings = _webScraper
@@ -218,12 +161,8 @@ class AsuraScans implements ManhwaSource {
           .toList();
 
       // ─── Get Manga Urls ──────────────────
-      final resultMangaUrls = _webScraper
-          .getElement('div.bsx > a', ['href'])
-          .map(
-            (e) => e['attributes']['href'].toString(),
-          )
-          .toList();
+      final resultMangaUrls =
+          _webScraper.getElementAttributeUnwrapString('div.bsx > a', 'href');
 
       // ─── Into Mangasearchresult List ─────