+++ lib/core/manhwa_sites/flame_scans.dart
@@ -1,4 +1,7 @@
+import 'dart:developer';
+
 import 'package:manga_reader/core/utils.dart';
+import 'package:manga_reader/core/webscraper_extension.dart';
 import 'package:web_scraper/web_scraper.dart';
 
 import '../core_types.dart';
@@ -13,14 +16,8 @@ class FlameScans implements ManhwaSource {
 
     if (await _webScraper.loadWebPage(chapterUrlEndpoint)) {
       // get images, then convert them to string and convert map to list
-      final allImages = _webScraper
-          .getElement('div#readerarea > p > img', ['src'])
-          .map(
-            (e) => e['attributes']['src'].toString(),
-          )
-          .toList();
-
-      return allImages;
+      return _webScraper.getElementAttributeUnwrapString(
+          'div#readerarea > p > img', 'src');
     }
 
     return [];
@@ -35,25 +32,21 @@ class FlameScans implements ManhwaSource {
       // ─── Get Title ───────────────────────────────────────
 
       final mangaTitle =
-          _webScraper.getElementTitle('div.titles > h1.entry-title')[0];
+          _webScraper.getFirstElementTitle('div.titles > h1.entry-title');
 
       // ─── Get Description ─────────────────────────────────
 
-      String mangaDescription = _webScraper
-          .getElementTitle('div.summary > div.wd-full > div.entry-content > p')
-          .first;
+      String mangaDescription = _webScraper.getFirstElementTitle(
+          'div.summary > div.wd-full > div.entry-content > p');
 
       // ─── Get Cover Url ───────────────────────────────────
 
-      final mangaCoverUrl = _webScraper
-          .getElement('div.thumb-half > div.thumb > img', ['src'])
-          .map(
-            (e) => e['attributes']['src'].toString(),
-          )
-          .first;
+      final mangaCoverUrl = _webScraper.getFirstElementAttribute(
+          'div.thumb-half > div.thumb > img', 'src');
 
       // ─── Get Status ──────────────────────────────────────
 
+      // cant use getFirstElementTitle here as the index is 1 😥
       final mangaStatus = MangaStatus.parse(
         _webScraper.getElementTitle('div.tsinfo > div.imptdt > i')[1],
       );
@@ -62,9 +55,8 @@ class FlameScans implements ManhwaSource {
 
       // Get Manga rating then parse it to double
       final mangaRating = double.parse(
-        _webScraper
-            .getElementTitle('div.extra-info > div.mobile-rt > div.numscore')
-            .first,
+        _webScraper.getFirstElementTitle(
+            'div.extra-info > div.mobile-rt > div.numscore'),
       );
 
       // ─── Get Released At ─────────────────────────────────
@@ -73,13 +65,8 @@ class FlameScans implements ManhwaSource {
       // 2. convert it to string
       // 3. parse it into `DateTime` object
       final mangaReleasedOn = DateTime.parse(
-        _webScraper
-            .getElement(
-              'div.imptdt > i > time',
-              ['datetime'],
-            )
-            .first['attributes']['datetime']
-            .toString(),
+        _webScraper.getFirstElementAttribute(
+            'div.imptdt > i > time', 'datetime'),
       );
 
       // ─── Get Tags ────────────────────────────────────────
@@ -91,8 +78,10 @@ class FlameScans implements ManhwaSource {
       // ─── Get Chapters ────────────────────────────────────
 
       // Get all chapter numbers
-      final mangaChapterTitless =
-          _webScraper.getElementTitle('span.chapternum').map((e) => removeChapterFromString(e)).toList();
+      final mangaChapterTitles = _webScraper
+          .getElementTitle('span.chapternum')
+          .map((e) => removeChapterFromString(e))
+          .toList();
 
       // Get all released on dates
       final mangaChapterReleasedOn = _webScraper
@@ -103,12 +92,8 @@ class FlameScans implements ManhwaSource {
           .toList();
 
       // Get all chapter urls
-      final mangaChapterUrls = _webScraper
-          .getElement('div#chapterlist > ul > li > a', ['href'])
-          .map(
-            (e) => e['attributes']['href'].toString(),
-          )
-          .toList();
+      final mangaChapterUrls = _webScraper.getElementAttributeUnwrapString(
+          'div#chapterlist > ul > li > a', 'href');
 
       // Comine all into list of MangaChapterDate
       List<MangaChapterData> mangaChapters = [];
@@ -116,7 +101,7 @@ class FlameScans implements ManhwaSource {
       for (var i = 0; i < mangaChapterUrls.length; i++) {
         mangaChapters.add(
           MangaChapterData(
-            mangaChapterTitless[i],
+            mangaChapterTitles[i],
             mangaChapterReleasedOn[i],
             mangaChapterUrls[i],
           ),
@@ -142,88 +127,40 @@ class FlameScans implements ManhwaSource {
 
   @override
   Future<List<MangaSearchResult>> popular({int page = 1}) async {
-    if (await _webScraper
-        .loadWebPage('/ygd/series/?type=manhwa&order=popular&page=$page')) {
-      // Get cover urls
-      final resultCoverUrls = _webScraper
-          .getElement(
-              'img.ts-post-image.wp-post-image.attachment-medium.size-medium',
-              ['src'])
-          .map((e) => e['attributes']['src'].toString())
-          .toList();
+    final targetEndpoint = '/ygd/series/?type=manhwa&order=popular&page=$page';
 
-      // get titles and get manga url
-      final resultTitleAndMangaUrl =
-          _webScraper.getElement('div.bsx > a', ['href', 'title']);
-
-      final resultTitles = resultTitleAndMangaUrl
-          .map((e) => e['attributes']['title'].toString())
-          .toList();
-
-      final resulMangaUrls = resultTitleAndMangaUrl
-          .map((e) => e['attributes']['href'].toString())
-          .toList();
-
-      // get rating
-      final resultRatings = _webScraper
-          .getElementTitle('div.mobile-rt > div.numscore')
-          .map(
-            (e) => double.parse(e),
-          )
-          .toList();
-
-      // get status
-      final resultStatuses = _webScraper
-          .getElementTitle('div.status > i')
-          .map(
-            (e) => MangaStatus.parse(e),
-          )
-          .toList();
-
-      List<MangaSearchResult> results = [];
-      // combine into list of MangaSearchResult
-      for (int i = 0; i < resulMangaUrls.length; i++) {
-        results.add(
-          MangaSearchResult(
-            resultCoverUrls[i],
-            resultTitles[i],
-            '',
-            resultRatings[i],
-            resulMangaUrls[i],
-            resultStatuses[i],
-          ),
-        );
-      }
-
-      return results;
-    }
-
-    return [];
+    return await _makeSearch(targetEndpoint);
   }
 
   @override
   Future<List<MangaSearchResult>> search(String query) async {
-    final formatedQuery = query.replaceAll(RegExp(r' '), '+').toLowerCase();
+    final formattedQuery = query.replaceAll(RegExp(r' '), '+').toLowerCase();
 
-    if (await _webScraper.loadWebPage('/ygd/?s=$formatedQuery')) {
+    return await _makeSearch('/ygd/?s=$formattedQuery');
+  }
+
+  Future<List<MangaSearchResult>> _makeSearch(String targetEndpoint) async {
+    if (await _webScraper.loadWebPage(targetEndpoint)) {
       // Get cover urls
-      final resultCoverUrls = _webScraper
-          .getElement(
-              'img.ts-post-image.wp-post-image.attachment-medium.size-medium',
-              ['src'])
-          .map((e) => e['attributes']['src'].toString())
-          .toList();
+      final resultCoverUrls = _webScraper.getElementAttributeUnwrapString(
+        'img.ts-post-image.wp-post-image.attachment-medium.size-medium',
+        'src',
+      );
 
       // get titles and get manga url
       final resultTitleAndMangaUrl =
           _webScraper.getElement('div.bsx > a', ['href', 'title']);
 
       final resultTitles = resultTitleAndMangaUrl
-          .map((e) => e['attributes']['title'].toString())
+          .map(
+            (e) => e['attributes']['title'].toString(),
+          )
           .toList();
 
       final resulMangaUrls = resultTitleAndMangaUrl
-          .map((e) => e['attributes']['href'].toString())
+          .map(
+            (e) => e['attributes']['href'].toString(),
+          )
           .toList();
 
       // get rating
@@ -249,7 +186,7 @@ class FlameScans implements ManhwaSource {
           MangaSearchResult(
             resultCoverUrls[i],
             resultTitles[i],
-            '',
+            null,
             resultRatings[i],
             resulMangaUrls[i],
             resultStatuses[i],