+++ lib/core/manhwa_sites/luminous_scans.dart
@@ -1,4 +1,5 @@
 import 'package:manga_reader/core/utils.dart';
+import 'package:manga_reader/core/webscraper_extension.dart';
 import 'package:web_scraper/web_scraper.dart';
 
 import '../core_types.dart';
@@ -12,14 +13,8 @@ class LuminousScans implements ManhwaSource {
         chapterUrl.replaceAll(RegExp(r'https://luminousscans.com'), '');
 
     if (await _webScraper.loadWebPage(chapterUrlEndpoint)) {
-      final allImages = _webScraper
-          .getElement('img.alignnone.size-full', ['src'])
-          .map(
-            (e) => e['attributes']['src'].toString(),
-          )
-          .toList();
-
-      return allImages;
+      return _webScraper.getElementAttributeUnwrapString(
+          'img.alignnone.size-full', 'src');
     }
 
     return [];
@@ -34,7 +29,7 @@ class LuminousScans implements ManhwaSource {
       // ─── Get Title ───────────────────────────────────────
 
       final title =
-          _webScraper.getElementTitle('div#titlemove > h1.entry-title').first;
+          _webScraper.getFirstElementTitle('div#titlemove > h1.entry-title');
 
       // ─── Get Tags ────────────────────────────────────────
 
@@ -49,21 +44,21 @@ class LuminousScans implements ManhwaSource {
 
       // ─── Get Cover Url ───────────────────────────────────
 
-      final coverUrl = _webScraper
-          .getElement('div.thumb > img', ['src']).first['attributes']['src'];
+      final coverUrl =
+          _webScraper.getFirstElementAttribute('div.thumb > img', 'src');
 
       // ─── Get Rating ──────────────────────────────────────
 
       // 1. get rating as string
       // 2. convert string to double
       final rating = double.parse(
-        _webScraper.getElementTitle('div.rating-prc > div.num').first,
+        _webScraper.getFirstElementTitle('div.rating-prc > div.num'),
       );
 
       // ─── Get Status ──────────────────────────────────────
 
       final status = MangaStatus.parse(
-        _webScraper.getElementTitle('div.tsinfo > div.imptdt > i').first,
+        _webScraper.getFirstElementTitle('div.tsinfo > div.imptdt > i'),
       );
 
       // ─── Get Year Released ───────────────────────────────
@@ -72,10 +67,8 @@ class LuminousScans implements ManhwaSource {
       // 2. convert it to string
       // 3. parse it into `DateTime` object
       final dateReleased = DateTime.parse(
-        _webScraper
-            .getElement('div.tsinfo > div.imptdt > i > time', ['datetime'])
-            .first['attributes']['datetime']
-            .toString(),
+        _webScraper.getFirstElementAttribute(
+            'div.tsinfo > div.imptdt > i > time', 'datetime'),
       );
 
       // ─── Get Chapters ────────────────────────────────────
@@ -85,9 +78,12 @@ class LuminousScans implements ManhwaSource {
       // 1. get chapter title `Chapter x`
       // 2. remove `Chapter `
       // 3. convert chapter number from string to double
-      final chapterTitles = _webScraper.getElementTitle(
-        'div#chapterlist > ul > li > div.chbox > div.eph-num > a > span.chapternum',
-      ).map((e) => removeChapterFromString(e)).toList();
+      final chapterTitles = _webScraper
+          .getElementTitle(
+            'div#chapterlist > ul > li > div.chbox > div.eph-num > a > span.chapternum',
+          )
+          .map((e) => removeChapterFromString(e))
+          .toList();
 
       // get chapter release date and convert them into `DateTime` objects
       final chapterReleasedOn = _webScraper
@@ -97,13 +93,8 @@ class LuminousScans implements ManhwaSource {
           .map((e) => altDateFormat.parse(e))
           .toList();
 
-      final chapterUrls = _webScraper
-          .getElement(
-            'div#chapterlist > ul > li > div.chbox > div.eph-num > a',
-            ['href'],
-          )
-          .map((e) => e['attributes']['href'].toString())
-          .toList();
+      final chapterUrls = _webScraper.getElementAttributeUnwrapString(
+          'div#chapterlist > ul > li > div.chbox > div.eph-num > a', 'href');
 
       // add chapter object to list
       for (var i = 0; i < chapterUrls.length; i++) {
@@ -135,68 +126,20 @@ class LuminousScans implements ManhwaSource {
 
   @override
   Future<List<MangaSearchResult>> popular({int page = 1}) async {
-    if (await _webScraper
-        .loadWebPage('/series?status=&type=&order=popular&page=$page')) {
-      // Get manga ruls and titles
-      final anchorTag =
-          _webScraper.getElement('div.bsx > a', ['href', 'title']);
+    final targetEndpoint = '/series?status=&type=&order=popular&page=$page';
 
-      final mangaUrls = anchorTag
-          .map(
-            (e) => e['attributes']['href'].toString(),
-          )
-          .toList();
-      final mangaTitles = anchorTag
-          .map(
-            (e) => e['attributes']['title'].toString(),
-          )
-          .toList();
-
-      // Get cover images
-      final coverUrls = _webScraper
-          .getElement('div.bsx > a > div.limit > img.wp-post-image', ['src'])
-          .map(
-            (e) => e['attributes']['src'].toString(),
-          )
-          .toList();
-
-      // Get latest Chapters
-      // gets the `Chapter xx` text then removes the `Chapter `, then converts the number to double
-      final latestChapterTitles = _webScraper
-          .getElementTitle('div.bsx > a > div.bigor > div.adds > div.epxs').map((e) => removeChapterFromString(e)).toList();
-
-      // Get Ratings
-      final ratings = _webScraper
-          .getElementTitle(
-              'div.bsx > a > div.bigor > div.adds > div.rt > div.rating > div.numscore')
-          .map((e) => double.parse(e))
-          .toList();
-
-      List<MangaSearchResult> results = [];
-
-      for (int i = 0; i < mangaUrls.length; i++) {
-        results.add(
-          MangaSearchResult(
-            coverUrls[i],
-            mangaTitles[i],
-            latestChapterTitles[i],
-            ratings[i],
-            mangaUrls[i],
-            MangaStatus.none,
-          ),
-        );
-      }
-      return results;
-    }
-
-    return [];
+    return await _makeSearch(targetEndpoint);
   }
 
   @override
   Future<List<MangaSearchResult>> search(String query) async {
     final formatedQuery = query.toLowerCase().replaceAll(RegExp(r' '), '-');
 
-    if (await _webScraper.loadWebPage('/?s=$formatedQuery')) {
+    return await _makeSearch('/?s=$formatedQuery');
+  }
+
+  Future<List<MangaSearchResult>> _makeSearch(String targetEndpoint) async {
+    if (await _webScraper.loadWebPage(targetEndpoint)) {
       // Get manga ruls and titles
       final anchorTag =
           _webScraper.getElement('div.bsx > a', ['href', 'title']);
@@ -213,17 +156,15 @@ class LuminousScans implements ManhwaSource {
           .toList();
 
       // Get cover images
-      final coverUrls = _webScraper
-          .getElement('div.bsx > a > div.limit > img.ts-post-image', ['src'])
-          .map(
-            (e) => e['attributes']['src'].toString(),
-          )
-          .toList();
+      final coverUrls = _webScraper.getElementAttributeUnwrapString(
+          'div.bsx > a > div.limit > img.wp-post-image', 'src');
 
       // Get latest Chapters
       // gets the `Chapter xx` text then removes the `Chapter `, then converts the number to double
       final latestChapterTitles = _webScraper
-          .getElementTitle('div.bsx > a > div.bigor > div.adds > div.epxs').map((e) => removeChapterFromString(e)).toList();
+          .getElementTitle('div.bsx > a > div.bigor > div.adds > div.epxs')
+          .map((e) => removeChapterFromString(e))
+          .toList();
 
       // Get Ratings
       final ratings = _webScraper
@@ -237,16 +178,15 @@ class LuminousScans implements ManhwaSource {
       for (int i = 0; i < mangaUrls.length; i++) {
         results.add(
           MangaSearchResult(
-            mangaUrls[i],
+            coverUrls[i],
             mangaTitles[i],
             latestChapterTitles[i],
             ratings[i],
-            coverUrls[i],
+            mangaUrls[i],
             MangaStatus.none,
           ),
         );
       }
-
       return results;
     }