+++ changelog.md
@@ -6,3 +6,8 @@ Commit 5:
 - refactored all `*chapterTitle` attributes in the `manga_types.dart`
 - renamed `manga_types.dart` to `core_types.dart`
 - ETC
+
+Commit 6:
+- Make `latestChapterNumber` in `MangaSearchResult` nullable and refactor accordingly
+- Make `description` in nullable and refactor accordingly
+- add extensions to `WebScraper` and refactor all manhwa sites